requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
boto3==1.34.0
python-multipart==0.0.6
spacy==3.7.2
openai==1.3.7
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
httpx==0.25.2

# Optional: for testing
pytest==7.4.3
